Develop

+++ src/services/deleteProducts.service.js
@@ -3,6 +3,17 @@ import database from "../database";
 const deleteProductsService = async (idProduct) => {
 
     try {
+
+        const idProducts = await database.query(
+            "SELECT id FROM products"
+        );
+
+        const idProductExists = idProducts.rows.find((row) => row.id === idProduct);
+
+        if(idProductExists === undefined){
+            throw ("Este produto não existe.")
+        }
+
         const res = await database.query(
             "DELETE FROM products WHERE name = ($1) RETURNING *",
             [idProduct]
